requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-vision==3.7.0
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
